Update the styling of the site templates
This change continues refining the style of the site pages so that they
take on an ever more professional appearance. The most amount of work
has gone in to the page header and then, to a lesser extent, the footer.
Previously, they were quite flat, but now they have a better appearance,
imho. I think that it can often be the little touches that make an app
really shine.