Fix up the header structure of the site
This change fixes up what I've been seeing as a problem with the site's
structure. That is, that the main link at the top of the site shouldn't
be the top-level header. Rather, the main header on a given page should
be. I don't know why I took the original approach, but I'm happy that
this change corrects the situation.