Update the icons in the site's forms
The one that was originally included in the forms was just a
placeholder. This change replaces it, where applicable, with ones that
make sense in the individual contexts. It also updates the relevant
styling so that it maintains a professional appearance. Credit for the
icons will be provided/attributed in the accompanying tutorial.