Rework the application's entity classes
This change refactors the User entity to have a many-to-one relationship
with a new suburb bin collection entity. This change will allow for
storing the collection date for each suburb (it's a simplistic model,
and assumes just one town with no suburb duplication) and to link users
to their bin collection days by suburb.