Refactored the SubscribeProcessorHandler
There are a series of changes in this commit tasked with making the code
smaller and cleaner, and easier to maintain. There are also changes
required to complete implementing all of the necessary functionality.

It adds an OpenStreetMap and EntityManager object as constructor
parameters, which are required to interact with the external services
the code needs. The SubscribeProcessorHandlerFactory is updated
accordingly.

It extracts the FlashMessenger code out into a trait, as the same code
is used in the UnsubscribeProcessorHandler.

It finishes updating the process method, handling cases where the
supplied data is both valid and invalid.

The change also adds a small class to flatten error messages. The
application uses laminas-validator and laminas-inputfilter to handle
validating the form input.

This works really well, but the error message format is a little
challenging, initially, to work with for rendering that information in a
template, such as the Twig templates which the app uses.

So this change adds a small class, ErrorsListFlattener that can flatten
the array of error messages into an associative array where the key is
the field and the value is an imploded list of all of the validation
messages for that field.