Refactor the SubscribeFormHandler's Twig template
This is a pretty reasonable refactor of the template.

It removes the original approach to rendering form errors, replacing it
with an inline approach, where the field and its label are highlighted
if there is a problem with that field. Any form validation error
messages are also printed next to the label.

In addition, existing form data is set in the respective form fields, if
present, allowing for the form to be filled out more quickly and easily.